Move configuration view into `frontend::configuration` module
